+++ ninja-core/src/site/markdown/developer/changelog.md
@@ -1,3 +1,13 @@
+ * 2016-01-19 SwissKnife converters replaced by usage of ParamParsers in BodyPostEngine (jlannoy)
+ * 2016-01-19 Support added for any enumeration value (registration deprecated) and custom ParamParser (jlannoy)
+
+Version 5.3.1
+=============
+
+ * 2016-01-29 Security: session and flash data save to Context instead of Result (jjlauer)
+ * 2016-01-16 Module.java of Ninja applications can now get NinjaProperties injected optionally. (jlannoy)
+ * 2015-12-01 Added more documentation about Freemarker templates. (jlannoy)
+
 Version 5.3.0
 =============